Add { target: 'node' } to module.exports for Production
This solves a problem using Webpack that calls on modules like fs and throws an error.

Per Webpack, "using node webpack will compile for usage in a Node.js-like environment (uses Node.js require to load chunks and not touch any built in modules like fs or path)."

A group of similar errors that was being thrown resembled the following:

"Module not found: Error: Can't resolve 'fs' in <VARIOUS FILE PATHS>"

AND

"There are multiple modules with names that only differ in casing.
This can lead to unexpected behavior when compiling on a filesystem with other case-semantic."